Feature/18 add photos to events

+++ Erasmus.Web/Controllers/ParticipantController.cs
@@ -1,16 +1,10 @@
 ﻿using AspNetCoreHero.ToastNotification.Abstractions;
 using Erasmus.Domain.Domain;
-using Erasmus.Domain.DomainModels;
 using Erasmus.Domain.DTO;
-using Erasmus.Repository.Implementation;
 using Erasmus.Repository.Interface;
 using Erasmus.Service.Interface;
-using GemBox.Document;
-using GemBox.Pdf;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.FileProviders;
 using Syncfusion.DocIO;
 using Syncfusion.DocIO.DLS;
 using Syncfusion.Pdf.Parsing;
@@ -25,16 +19,18 @@ namespace Erasmus.Web.Controllers
         private readonly INonGovProjectService _projectService;
         private readonly IUploadedFileRepository _uploadedFileRepository;
         private readonly IUserRepository _userRepository;
-        private readonly IParticipantsRepository _participantRepository;
+        private readonly IParticipantApplicationService _participantApplicationService;
+        private readonly IParticipantService _participantService;
         private readonly INotyfService _notyfService;
         public ParticipantController(INonGovProjectService projectService, IUploadedFileRepository uploadedFileRepository, IUserRepository userRepository,
-            IParticipantsRepository participantRepository, INotyfService notyfService)
+            IParticipantService participantService, INotyfService notyfService, IParticipantApplicationService participantApplicationService)
         {
             _notyfService = notyfService;
             _projectService = projectService;
             _uploadedFileRepository = uploadedFileRepository;
             _userRepository = userRepository;
-            _participantRepository = participantRepository;
+            _participantApplicationService = participantApplicationService;
+            _participantService = participantService;
             GemBox.Document.ComponentInfo.SetLicense("FREE-LIMITED-KEY");
             GemBox.Pdf.ComponentInfo.SetLicense("FREE-LIMITED-KEY");
         }
@@ -48,6 +44,7 @@ namespace Erasmus.Web.Controllers
         {
             var project = _projectService.Get(eventId);
             var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var application = _participantApplicationService.GetForParticipantAndProject(userId, eventId);
             var uploadedCv = _uploadedFileRepository.UploadedCvForEvent(userId, eventId);
             var uploadedMotivation = _uploadedFileRepository.UploadedMotivationalLetterForEvent(userId, eventId);
             if (project != null)
@@ -57,9 +54,18 @@ namespace Erasmus.Web.Controllers
                     ProjectId = project.Id,
                     Project = project,
                     UploadedCV = uploadedCv,
-                    UploadedMotivation = uploadedMotivation
+                    UploadedMotivation = uploadedMotivation,
+                    ParticipantId = userId
                 };
-                //model.UploadedFilesForUser = _uploadedFileRepository.GetFilesForUserAndEvent(userId, eventId);
+
+                if(application != null)
+                {
+                    model.ReviewStatus = application.ReviewStatus;
+                }
+                else
+                {
+                    model.ReviewStatus = ApplicationStatus.NotCompleted;
+                }
                 return View(model);
             }
             else
@@ -74,7 +80,7 @@ namespace Erasmus.Web.Controllers
             var userId = User.FindFirstValue(ClaimTypes.NameIdentifier).ToString();
             var project = _projectService.Get(model.ProjectId);
             model.Project = project;
-            var participant = _participantRepository.Get(userId);
+            var participant = _participantService.Get(userId);
             try
             {
                 if (model.CV != null)
@@ -122,7 +128,7 @@ namespace Erasmus.Web.Controllers
                 }
                 if(model.CV != null || model.UploadedCV != null)
                      _notyfService.Success("Files uploaded!");
-                return RedirectToAction("Details", "NonGovProjects", new { id = model.ProjectId });
+                return RedirectToAction("UploadFiles", "Participant", new { eventId = model.ProjectId });
 
             }
             catch
@@ -205,14 +211,51 @@ namespace Erasmus.Web.Controllers
             // delete the record
             _uploadedFileRepository.Delete(id);
             // delete the actual file
-            System.IO.File.Delete(file.PathOnDisk);
+            var participantId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var application = _participantApplicationService.GetForParticipantAndProject(participantId, file.ProjectId);
+            if(application != null)
+                application.ReviewStatus = ApplicationStatus.NotCompleted;
+            _participantApplicationService.Update(application);
+
+            if(!IsFileLocked(file.PathOnDisk))
+                System.IO.File.Delete(file.PathOnDisk);
             _notyfService.Success("File deleted");
-            return RedirectToAction("UploadFiles", "Participant", new { eventId = id});
+            return RedirectToAction("UploadFiles", "Participant", new { eventId = file.ProjectId});
         }
 
         public IActionResult Profile()
         {
             return View();
+        }
+
+        public IActionResult Apply(ApplyToEventDto model)
+        {
+            // create application in db
+            _participantService.Apply(model.ParticipantId, model.ProjectId);
+            _notyfService.Success("Your application for the event is successful!");
+            return RedirectToAction("UploadFiles", new { eventId = model.ProjectId});
+        }
+
+        protected virtual bool IsFileLocked(string path)
+        {
+            try
+            {
+                using (FileStream stream = System.IO.File.Open(path, FileMode.Open))
+                {
+                    stream.Close();
+                }
+            }
+            catch (IOException)
+            {
+                //the file is unavailable because it is:
+                //still being written to
+                //or being processed by another thread
+                //or does not exist (has already been processed)
+                return true;
+            }
+
+            //file is not locked
+            return false;
         }
     }
 }
